@@ -18,15 +18,15 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.18.0 <3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
   http:
-  qr_flutter: ^3.2.0
-  barcode_scan: ^2.0.1
-  toast:
+  qr_flutter: ^4.1.0
+  barcode_scan2: ^4.2.4
+  toast: 
 
 
   # The following adds the Cupertino Icons font to your application.
@@ -51,7 +51,8 @@ flutter:
   uses-material-design: true
 
   # To add assets to your application, add an assets section, like this:
-  # assets:
+  assets:
+   - assets\TrashTagBg.png
   #   - images/a_dot_burr.jpeg
   #   - images/a_dot_ham.jpeg
 
