@@ -15,7 +15,7 @@ class MyApp extends StatelessWidget {
       title: 'TrashTag',
       debugShowCheckedModeBanner: false,
       theme: ThemeData(
-        primarySwatch: Colors.blue,
+        primarySwatch: Colors.green,
       ),
       home: TrashTagApp(),
     );
@@ -23,47 +23,65 @@ class MyApp extends StatelessWidget {
 }
 
 class TrashTagApp extends StatelessWidget {
-  const TrashTagApp({Key key}) : super(key: key);
+  const TrashTagApp({super.key});
 
   @override
   Widget build(BuildContext context) {
     return Scaffold(
-      appBar: AppBar(
-        title: Text("TrashTag"),
-      ),
-      body: Center(
-        child: Container(
-          padding: EdgeInsets.all(20),
-          child: Column(
-            children: [
-              RaisedButton(
-                onPressed: () {
-                  Navigator.of(context).push(
-                      MaterialPageRoute(builder: (contet) => LoginScreen()));
-                },
-                child: Text("Login"),
-              ),
-              RaisedButton(
-                onPressed: () {
-                  Navigator.of(context).push(
-                      MaterialPageRoute(builder: (contet) => RegisterScreen()));
-                },
-                child: Text("Register"),
-              )
-              // SizedBox(
-              //   height: 20,
-              // ),
-              // RaisedButton(
-              //   onPressed: () {
-              //     Navigator.of(context)
-              //         .push(MaterialPageRoute(builder: (contet) => ScanPage()));
-              //   },
-              //   child: Text("QR Test"),
-              // )
-            ],
-          ),
+        appBar: AppBar(
+          title: Text("TrashTag"),
         ),
-      ),
-    );
+        body: Container(
+          decoration: BoxDecoration(
+            image: DecorationImage(
+              image: AssetImage("assets/TrashTagBg.png"),
+              fit: BoxFit.cover,
+            ),
+          ),
+          child: Center(
+            child: Container(
+              padding: EdgeInsets.all(20),
+              child: Center(
+                child: Column(
+                  mainAxisAlignment: MainAxisAlignment.center,
+                  children: [
+                    Row(
+                      mainAxisAlignment: MainAxisAlignment.center,
+                      children: [
+                        ElevatedButton(
+                          onPressed: () {
+                            Navigator.of(context).push(MaterialPageRoute(
+                                builder: (contet) => LoginScreen()));
+                          },
+                          child: Text("Login"),
+                        ),
+                        SizedBox(
+                          width: 30,
+                        ),
+                        ElevatedButton(
+                          onPressed: () {
+                            Navigator.of(context).push(MaterialPageRoute(
+                                builder: (contet) => RegisterScreen()));
+                          },
+                          child: Text("Register"),
+                        ),
+                      ],
+                    ),
+                    SizedBox(
+                      height: 20,
+                    ),
+                    ElevatedButton(
+                      onPressed: () {
+                        Navigator.of(context).push(
+                            MaterialPageRoute(builder: (contet) => ScanPage()));
+                      },
+                      child: Text("QR Test"),
+                    )
+                  ],
+                ),
+              ),
+            ),
+          ),
+        ));
   }
 }
