@@ -3,7 +3,7 @@ import 'package:trashtag/login.dart';
 import 'package:trashtag/server.dart';
 
 class RegisterScreen extends StatefulWidget {
-  const RegisterScreen({Key key}) : super(key: key);
+  const RegisterScreen({super.key});
 
   @override
   _RegisterScreenState createState() => _RegisterScreenState();
@@ -43,7 +43,7 @@ class _RegisterScreenState extends State<RegisterScreen> {
               SizedBox(
                 height: 20,
               ),
-              RaisedButton(
+              ElevatedButton(
                 child: Text("Register"),
                 onPressed: () {
                   print(uc.text);
